+++ inst/examples/gg.spatial.R
@@ -1,6 +1,7 @@
 \donttest{
   if (require(ggplot2, quietly = TRUE) &&
-      require(ggpolypath, quietly = TRUE)) {
+      require(ggpolypath, quietly = TRUE) &&
+      bru_safe_sp()) {
     # Load Gorilla data
 
     data("gorillas", package = "inlabru")
@@ -22,7 +23,7 @@
 
     data("mexdolphin", package = "inlabru")
 
-    # Plot the pantropiical survey boundary, ship transects and dolphin sightings
+    # Plot the pantropical survey boundary, ship transects and dolphin sightings
 
     ggplot() +
       gg(mexdolphin$ppoly) + # survey boundary as SpatialPolygon