+++ tests/testthat/test-lgcp_2d.R
@@ -69,7 +69,7 @@ test_that("2D LGCP fitting", {
 
   # test_that("2D LGCP fitting: predicted random field", {
   loc <- SpatialPoints(gorillas$mesh$loc[, c(1, 2)],
-    proj4string = INLA::inla.sp_get_crs(gorillas$nests)
+    proj4string = fm_CRS(gorillas$nests)
   )
   set.seed(123L)
   pr <- predict(fit, loc, ~mySmooth,